requirements.txt
pettingzoo>=1.24.0
supersuit>=3.9.0
stable-baselines3>=2.3.0
torch>=2.0.0
numpy>=1.24.0
yamlargparse>=1.26.0
pyyaml>=6.0.0
moviepy>=1.0.3
wandb>=0.17.0
